Split default navigation bar specs
REDMINE-20481
